tools: kdump: introduce arm64 kdump addon support patches for LSP
The kdump support for arm64 is still on the kernel mainling list.
Therefore patches are provided as an addon which can be applied on
top of Marvel LSP release.

Patches was ported from:
git: https://git.linaro.org/people/takahiro.akashi/linux-aarch64.git
branch: origin/lsk/kdump/for-v4.4

It is also worth mentioning that the patches can't be ported directly from
the mainline lists, since the gap between kernel v4.4 and mainline 4.10 is
big and the whole support touches a lot of subsystem including generic mm
and arm64. Because of that the amount of patches which are required before
applying the main arm64's  kexec/kdump patches is so big.

Change-Id: I7dd43301f2c53fdf6ef79b2fcd8323a9164296ee
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/37197
Reviewed-by: Nadav Haklai <nadavh@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Omri Itach <omrii@marvell.com>