fix: dts: a70x0: relocate A7040-DB specific IO expander interrupt to configuration D
IO expander interrupt is connected to MPP7 on
armada-7040-DB for setup D only,
while on setup A,B,C,E its connected to MPP62.
MPP62 is used for SD/MMC for setups A,B,C,E,
hence can't be used for IO expander interrupt.
Therefore these changes should be in D configuration
device tree and not in the 7040-db.dtsi.

Change-Id: Iac189c242c9ba2e5defc690fe52afec92d5283f1
Signed-off-by: Christine Gharzuzi <chrisg@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/38557
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Omri Itach <omrii@marvell.com>