fix: usb: gadget: u3d: send EP ready until mass storage thread is ready
During system suspend to RAM, the kernel modules will be resumed at
first then re-start all the kernel threads, which leads to issue for
USB3 device mode. The USB3 device driver will start negotiate with host
when mass storage thread is not resumed, which leads to failure and
finally the host will set EP to suspend state, thus the USB3 device
can not work even after the mass storage thread is resumed.
This patch modify the u3d driver that do not send state ready to host
when the function report USB_GADGET_DELAYED_STATUS state, which is used
to delay the data/status stages of the control transfer till function
are ready.
In routine mvc2_handle_setup(), the USB_GADGET_DELAYED_STATUS will be
checked so that do not continue the transfer if it is set.
After the kernel USB mass storage thread is resumed, it will handle
the setup, then send the request to notify lower layer that the delayed
status is cleared by set the request packet length to 0.
In Marvell USB3 device driver, once it receives the request whose length
is 0, it knows the delay status is cleared, then it will notify the host
that EP is ready(by calling mvc2_send_erdy) to continue the transfer.
JIRA ID: A3700-858 USB3 device do not resume when working with SATA disk
connected.

Change-Id: If3bb547c5ff7f8563ba002db7478a6d46bf961f3
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/38669
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>