pinctrl: armada-37xx: Add irqchip support
The Armada 37xx SoCs can handle interrupt through GPIO. However it can
only manage the edge ones.

The way the interrupt are managed is classical so we can use the generic
interrupt chip model.

The only unusual "feature" is that many interrupts are connected to the
parent interrupt controller. But we do not take advantage of this and
use the chained irq with all of them.

Backport gpio/pinctl patch [1/2] from linux mainline
Commit 2f227605394b ("pinctrl: armada-37xx: Add irqchip support").

Cherry-picked from branch linux-4.12-devel with a minor change that
private pinctrl info is not got from gpip chip data but got as a
container of gpio chip due to the different gpio irq framework
between 4.12 and 4.4.52.

Change-Id: Icdb13888d463c7874c72e621cb6918bcfecc8c15
Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Reviewed-on: http://vgitil04.il.marvell.com:8080/42541
Reviewed-by: Victor Gu <xigu@marvell.com>
Tested-by: Victor Gu <xigu@marvell.com>
Signed-off-by: Ken Ma <make@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/42780
Tested-by: iSoC Platform CI <ykjenk@marvell.com>