fix: dts: a3700: add comphy in PCIe node for community board
The PCIe comphy should be initialized either during cold boot
and PM resume if PCIe is enabled.
The missing of the comphy in community PCIe node will lead to issues
such as suspend to RAM failure during resume. It is because the PCIe
will not work if the PCIe comphy is not initialized, and even worse the
CPU will get stuck when access PCIe controller register if the PCIe
comphy is not initialized.
This patch adds the comphy to be referenced by PCIe node for
community board to fix this issue.
JIRA A3700-1188.

Change-Id: Ic9b16c23e93406a2754dd3e0f8d4125f638ee310
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/39609
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Evan Wang <xswang@marvell.com>
Reviewed-by: Terry Zhou <bjzhou@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/39671