fix: usb: gadget: u3d: remove HW reset from VBUS session
During the long time test, for SoC of armada3700, it is
found that the HW reset could lead to CPU stuck in vbus
session stage in USB3 device mode, and this issue is fixed
by removing the HW reset from vbus session routine
according to testing.

The patch removes HW reset from U3D vbus session routine
only for armada3700, which is just a WA to fix the issue.

Change-Id: Idf23264fc06725a813afaeee0795ea818e9e8c79
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/40218
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>