fix: net: mv88e6xxx: stu read and write for 88e6390.
Modified read and write of stu data as per functional specification.
Refer to Global 1, register 7 and register 8 of peridot and other
devices. These registers specify, ingress and egress membership and
egress tagging of each port. For all devices expect peridot, width of
the membership tag for each port is 4 bits wide in which two bits are
reserved. For peridot, width of membership tag for each port is two bits.

Change-Id: I8c35bb205d0b6d10636cf974835b6ee336d26658
Signed-off-by: Ravindra Reddy K <ravindra@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/46883
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Omri Itach <omrii@marvell.com>