net: mvpp2x: align aggregated queue struct to cache
This patch is FIX and Optimization.

The Aggregated TX queue descriptors are allocated and
referenced as array of descriptors/structures "per CPU".
The size of an entry in the array is 56B.
To avoid "cache line interference " on array's entry
the descriptor address and size should be be aligned to the
cache-line-size.
This alignment also improves the performance since 1 cache-line
used by CPU instead of 2 cache-lines.

Change-Id: I10f9ab1f2607bbc233fec69e7fe25a05cac3c4b8
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/48404
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Reviewed-by: Yuval Caduri <cyuval@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>